prepare_election/requirements.txt
geopandas
geopy
pandas
google-cloud-bigquery
google-cloud-storage
geocodio